:bug: Make gulp-cache tmp directory username-dependent.
Previous to this commit, there is a possibility that when one user on a
given system runs `gulp scss-lint`, followed by a _different_ user running
the same command, the gulp task will error because the second user doesn't
have permissions to delete /tmp/gulp-cache. To fix this, this commit
appends the current user's name to the temporary directory being used for
the cache. This effectively busts the cache for different users on the same
system.

Refs taigaio/taiga-front#1416.